Create FreeBibleVersionBible
Free Bible Version
The New Testament and Psalms in the Free Bible Version translation in American English
Copyright © 2018 Dr. Jonathan Gallagher
Language: English
Dialect: American
Translation by: Dr. Jonathan Gallagher

This translation is made available to you under the terms of the Creative Commons Attribution Share-Alike license 4.0.

You have permission to share and redistribute this Bible translation in any format and to make reasonable revisions and adaptations of this translation, provided that:

You include the above copyright and source information.
If you make any changes to the text, you must indicate that you did so in a way that makes it clear that the original licensor is not necessarily endorsing your changes.
If you redistribute this text, you must distribute your contributions under the same license as the original.
Pictures included with Scriptures and other documents on this site are licensed just for use with those Scriptures and documents. For other uses, please contact the respective copyright owners.

Note that in addition to the rules above, revising and adapting God's Word involves a great responsibility to be true to God's Word. See Revelation 22:18-19.

2025-09-10

Last updated 2025-09-10