Create WEBBible
The World English Bible British Edition with Deuterocanon
The World English Bible British Edition is not copyrighted. It is in the Public Domain. However "World English Bible" is a Trademark.

You may copy, publish, proclaim, distribute, redistribute, sell, give away, quote, memorize, read publicly, broadcast, transmit, share, back up, post on the Internet, print, reproduce, preach, teach from, and use the World English Bible as much as you want, and others may also do so. All we ask is that if you CHANGE the actual text of the World English Bible in any way, you not call the result the World English Bible any more. This is to avoid confusion, not to limit your freedom. The Holy Bible is God's Word. It belongs to God. He gave it to us freely, and we who have worked on this translation freely give it to you by dedicating it to the Public Domain.

Donations to help with the expenses of this project and thus help others have free access to the Word of God may be made to us, but are not required. Please see https://MLJohnson.org/partner/ for more about that.

For a list of what is different between the World English Bible and the World English Bible British Edition, besides the handling of God's Proper Name, see phrasetranslog.txt and phrasetranslist.txt.

2020 stable text edition

Last updated 2025-10-28