Create asvBible
The Bibles available on this page are FREE to use and distribute for non-commercial purposes. 

There is no need to ask us for permission.  Please share them with others.

Bibles that are in the public domain are free to use for any purpose, including commercial.

Please see the copyright statement on each Bible for details.

Note: We do not own the rights to any of these Bibles.  If you would like to use a given Bible for commercial purposes, you will need to obtain permission from the copyright holder directly.

https://www.biblesupersearch.com/bible-downloads/